requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
multidict==6.4.4
mlflow>=2.5.0
openai==1.84.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.5
pydantic-core==2.33.2
PyMuPDF==1.26.0
pytesseract==0.3.13
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
PyYAML>=6.0